requirements.txt
# FastAPI 및 웹 프레임워크  
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# 데이터베이스 
asyncpg>=0.29.0
redis>=5.0.0

# 데이터 검증
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 인증 및 보안 (필수)
PyJWT>=2.8.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4

# AI 및 LangChain 
openai>=1.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0

# 환경변수
python-dotenv>=1.0.0

# 크롤링 라이브러리
selenium>=4.16.0
beautifulsoup4>=4.12.0
requests>=2.31.0

traceloggerx==0.1.8
pytz>=2023.3